Created fields to store the first and the last nonzero coefficients
With that, the function Poly::degree becomes O(1), and the same happens to "find the first nonzero coefficient".
On the other hand, these fields have to be keeped up-to-date, so, function Poly::set now can be O(n), but its expected cost is O(1), since the polynomials are supposed to be dense.

I also created a field that tells if the coefficient is equal to zero or not.